Added snuggletex.extensions package hierarchy to contain "extra" or non-core stuff. The idea is to put samples and 3rd party helpers here whilst keeping the core of the project small and self-contained.
First extension added is a migration of the work done in Aardvark that uses JEuclid convert MathML to images. This enables us to generate legacy web pages quite nicely. Combined with the new down-conversion feature, this may be quite useful in some situations.

This first commit of this functionality is still a bit rough - some further tidying and documentation is needed! I'm also going to rename some stuff...